requirements.txt
build
debugpy
pyright
nbclassic
notebook
twine
openai
pillow
ipynbname
google-search-results
git+https://github.com/kiyoon/jupynium.nvim@v0.2.1
ipycanvas
orjson
wikipedia
